+++ src/index.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import App from "./App";
-import About from "./About";
-import "./App.css";
+import App from "./pages/Home";
+import About from "./pages/About";
+import "./index.css";
 import reportWebVitals from "./reportWebVitals";
 
 ReactDOM.render(